+++ src/tests/Morphant.Generator.UnitTests/ConstructionAndMembersUsageTests/StructuredResults/ConstructionLocalAndSwitch.cs
@@ -111,13 +111,7 @@ namespace TestCase
         {
             _ = global::Morphant.Option<global::TestCase.Destination>.None.TryGetValue(out global::TestCase.Destination? existing);
 
-            var name = source.FromConstruct();
-
-            return new global::TestCase.Destination(
-                name: name)
-            {
-                Name = source.FromMembers()
-            };
+            return __Construct(source, context);
         }
 
         private global::TestCase.Destination __Update(
@@ -147,16 +141,21 @@ namespace TestCase
                     }
                 }
             }
-            else
-            {
-                var name = source.FromConstruct();
 
-                return new global::TestCase.Destination(
-                    name: name)
-                {
-                    Name = source.FromMembers()
-                };
-            }
+            return __Construct(source, context);
+        }
+
+        private global::TestCase.Destination __Construct(
+            global::TestCase.Source source,
+            global::Morphant.Context.MappingContext context)
+        {
+            var name = source.FromConstruct();
+
+            return new global::TestCase.Destination(
+                name: name)
+            {
+                Name = source.FromMembers()
+            };
         }
     }
 }