+++ src/tests/Morphant.Generator.UnitTests/ConstructionAndMembersUsageTests/StructuredResults/StoredBooleanSwitch.cs
@@ -128,11 +128,7 @@ namespace TestCase
                 }
                 case false:
                 {
-                    return new global::TestCase.Destination(
-                        name: source.FromConstruct())
-                    {
-                        Name = source.FromMembers()
-                    };
+                    return __Create(source, context);
                 }
             }